requirements.txt
streamlit
python-dotenv
langchain
langchain-community
transformers
sentence-transformers
torch
faiss-cpu
unstructured
